+++ NEWS.md
@@ -1,4 +1,8 @@
-# version 0.2-1
+# version 0.3-0
+
+* use "native" R array-factor support
+
+* support for `PCICt` 360- and 365-day calendars; #29
 
 * remove import of `ncdf4` in favour of `RNetCDF`, now in line with practice in `ncmeta` package. 
 
@@ -6,7 +10,7 @@
 
 * allow missing trailing comma's when subsetting: `a[1,,]` and `a[1,]` now do the same.
 
-* move `rlang` to Imports: ; rewrite subset using rlang.
+* move `rlang` to Imports: ; rewrite `[` subset using rlang.
 
 * add conversion to and from `Spatial*` classes, including the gridded ones, taking care of `factor` variables